Organized classes into separate packages duke.task and duke.util
