+++ src/main/java/duke/task/Deadlines.java
@@ -1,6 +1,6 @@
-package duke;
+package duke.task;
 
-import duke.DateAndTimeFormatter;
+import duke.util.DateAndTimeFormatter;
 
 import java.time.LocalDate;
