+++ src/main/java/duke/util/DateAndTimeFormatter.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.util;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
